requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
apscheduler==3.10.4
requests==2.32.3
python-dotenv==1.0.1
